app/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
djangorestframework-simplejwt==5.2.2
#psycopg2>=2.8.6,<2.9
django-cors-headers==3.14.0